requirements.txt
django
django-cleanup
django-colorfield
django-polymorphic
Pillow